Fix to handle the mutually-recursive case
The revised "busy" logic was unable to handle a mutually-recursive macro like:

    #define ABC XYZ
    #define XYZ ABC

This change restores the ability to handle mutually recursive cases, and makes sure that we test that case.

The crux of the fix relies on splitting the environment and busy-macro tracking into more separate data structures.
Rather than the list of environments directly trakcing busy-ness via the chain of parent environments, an environment now stores a separate linked list of macros that should be considered busy in that environment.

Decoupling the two lists allows for an important change: when expanding an ordinary macro (either object-like or function-like, but *not* a function argument) the parent *environment* comes from the point where the macro was defined (this is required for lookup to make sense), but the parent list of *busy macros* comes from the current input stream at the point where expansion takes place. That change ensures that non-argument macros always properly "stack up" the list of busy macros.