Don't allocate a default space for a VK push-constant buffer
When a shader only uses `ParameterBlock`s plus a single buffer for root constants:

```hlsl
ParameterBlock<A> a;
ParameterBlock<B> b;

[[vk::push_constant]] cbuffer Stuff { ... }
```

we expect the push-constant buffer should not affect the `space` allocated to the parameter blocks (so `a` should get `space=0`).

This behavior wasn't being implemented correctly in `slang-parameter-binding.cpp`. There was logic to ignore certain resource kinds in entry-point parameter lists when computing whether a default space is needed, but the equivalent logic for the global scope only considered parameters that consuem whole register spaces/sets.

This change shuffles the code around and makes sure it considers a global push-constant buffer as *not* needing a default space/set.

Note that this change will have no impact on D3D targets, where `Stuff` above would always get put in `space0` because for D3D targets a push-constant buffer is no different from any other constant buffer in terms of register/space allocation.

One unrelated point that this change brings to mind is the `[[vk::push_constant]]` should ideally also be allowed to apply to an entry point (where it would modify the default/implicit constant buffer). In fact, it could be argued that push-constant allocation should be the *default* for (non-RT) entry point `uniform` parameters (while `[[vk::shader_record]]` should be the default for RT entry point `uniform` parameters).