Fix a reference-counting bug in one of the session creation routines. (#1230)
This is pretty straightforward, because we were calling `Session::init` (which can retain/release the session) on a `Session*` (no reference held).

The catch is that our current tests use the older form of the Slang API, while Falcor relies on the newer API, and so the recent change to our reference-counting logic introduced a regression that we didn't detect in testing.

This change just fixes the direct issue but doesn't address the gap in testing. A better long-term fix would be to fully define our "1.0" API, shift our testing to it, and layer the old API on top of it (to try and avoid regressions for client code).